+++ src/main/java/com/epam/ta/reportportal/core/item/UpdateTestItemHandlerImpl.java
@@ -44,6 +44,7 @@ import com.epam.ta.reportportal.ws.model.issue.Issue;
 import com.epam.ta.reportportal.ws.model.issue.IssueDefinition;
 import com.epam.ta.reportportal.ws.model.item.AddExternalIssueRQ;
 import com.epam.ta.reportportal.ws.model.item.UpdateTestItemRQ;
+import com.epam.ta.reportportal.ws.model.launch.Mode;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
@@ -176,7 +177,7 @@ public class UpdateTestItemHandlerImpl implements UpdateTestItemHandler {
 				testItem.setIssue(testItemIssue);
 
 				testItemRepository.save(testItem);
-				indexLogs(projectName, testItem);
+				indexLogs(projectName, testItem, launch.getMode());
 
 				testItem = statisticsFacadeFactory.getStatisticsFacade(project.getConfiguration().getStatisticsCalculationStrategy())
 						.updateIssueStatistics(testItem);
@@ -261,14 +262,26 @@ public class UpdateTestItemHandlerImpl implements UpdateTestItemHandler {
 	 * @param projectName Project name
 	 * @param testItem    Test item to reindex
 	 */
-	private void indexLogs(String projectName, TestItem testItem) {
-		if (testItem.getIssue().isIgnoreAnalyzer() || testItem.getIssue()
-				.getIssueType()
-				.equals(TestItemIssueType.TO_INVESTIGATE.getLocator())) {
-			logIndexer.cleanIndex(projectName, singletonList(testItem.getId()));
-		} else {
+	private void indexLogs(String projectName, TestItem testItem, Mode mode) {
+		if (canBeIndexed(testItem, mode)) {
 			logIndexer.indexLogs(testItem.getLaunchRef(), singletonList(testItem));
+		} else {
+			logIndexer.cleanIndex(projectName, singletonList(testItem.getId()));
 		}
+	}
+
+	/**
+	 * Test item could be indexed if it is not ignored for analyzer ,
+	 * issue is not TO_INVESTIAGE and launch mode is DEFAULT.
+	 *
+	 * @param testItem Test item
+	 * @param mode     Launch mode
+	 * @return True if can be indexed
+	 */
+	private boolean canBeIndexed(TestItem testItem, Mode mode) {
+		return (!testItem.getIssue().isIgnoreAnalyzer() || !testItem.getIssue()
+				.getIssueType()
+				.equals(TestItemIssueType.TO_INVESTIGATE.getLocator())) && mode.equals(Mode.DEFAULT);
 	}
 
 	/**